# src/main/resources/openapi/schema.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'VSA Projekt 2 schema'
  license:
    name: MIT
servers:
  - url: 'localhost:8080/api'
paths:
...
  /carparks/{id}/floors/{identifier}:
    get:
        summary: 'Get car park floor that belongs to car park with given id and has given identifier'
        operationId: getCarParkFloor
        parameters:
            - name: id
              description: 'Id of car park'
              required: true
              in: path
              schema:
                type: integer
                format: int64
            - name: identifier
              description: 'Identifier of the car park floor'
              required: true
              in: path
              schema:
                type: string
        responses:
            '200':
                description: 'Car park floor found and returned'
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CarParkFloor'
            '404':
                  description: 'Car park with given id was not found or car park floor with given identifier for given car park was not found'
    delete:
        summary: 'Delete car park floor for car park with given id with given identificator'
        operationId: deleteCarParkFloor
        parameters:
            - name: id
              description: 'Id of car park'
              required: true
              in: path
              schema:
                type: integer
                format: int64
            - name: identifier
              description: 'Identifier of the car park floor'
              required: true
              in: path
              schema:
                type: string
        responses:
            '204':
                description: 'Car park floor deleted'
            '404':
                description: 'Either car park with given id does not exist or there is no floor for the car park with given identificator'
...
  /parkingSpots/{id}:
      get:
            summary: 'Get parking spot by id'
            operationId: getParkingSpot
            parameters:
              - name: id
                in: path
                required: true
                description: 'Id of parking spot'
                schema:
                    type: integer
                    format: int64
            responses:
                '200':
                    description: 'Parking spot exists and is returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ParkingSpot'
                '404':
                    description: 'Parking spot does not exist'
      put:
            summary: 'Update parking spot'
            operationId: updateParkingSpot
            parameters:
              - name: id
                in: path
                required: true
                description: 'Id of parking spot'
                schema:
                    type: integer
                    format: int64
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ParkingSpot'
            responses:
                '201':
                    description: 'Parking spot updated and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ParkingSpot'
                '404':
                    description: 'Parking spot does not exist'
                '400':
                    description: 'Incorrectly formed request body'
      delete:
            summary: 'Delete parking spot'
            operationId: deleteParkingSpot
            parameters:
              - name: id
                in: path
                required: true
                description: 'Id of parking spot'
                schema:
                    type: integer
                    format: int64
            responses:
                '204':
                    description: 'Parking spot deleted'
                '404':
                    description: 'Parking spot does not exist'
...
components:
  schemas:
...
    CarParkFloor:
        type: object
        required:
            - id
            - identifier
            - carPark
        properties:
            id:
                type: integer
                format: int64
            identifier:
                type: string
            carPark:
                type: integer
                format: int64
            spots:
                type: array
                items:
                    typ: object
                    $ref: ParkingSpot
    ParkingSpot:
        type: object
        required:
            - id
            - identifier
            - carParkFloor
            - carPark
            - free
        properties:
            id:
                type: integer
                format: int64
            identifier:
                type: string
            carParkFloor:
                type: string
            carPark:
                type: integer
                format: int64
            type:
                type: object
                $ref: CarType  
            free:
                type: boolean
            reservations:
                type: array
                items:
                    $ref: ReservationDownFromParkingSpot
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
security:
  - BasicAuth: []